requirements.txt
mediapipe
opencv-python>=3.4.2
tensorflow>=2.3.0
tf-nightly>=2.5.0.dev
scikit-learn>=0.23.2 
matplotlib>=3.3.2 
numpy>=1.19.5
pandas>=1.1.5
scipy>=1.5.2
tensorflow-macos
tensorflow-metal
jupyter
pyobjc==11.0
pyautogui==0.9.54

